backend/requirements.txt
dj-rest-auth
djangorestframework
djoser
django-allauth
python-dotenv
psycopg2